Update README.md
This project solves the problem of time-consuming manual attendance and proxy attendance in classrooms. The system works by:

Pre-inserting student images into the Python script.

Requiring the student to scan their face upon entering the class.

Automatically marking their attendance as "Present" and recording the date and time in an Excel sheet if a face match is successful.